+++ packages/adapter-mongoose/tests/list-adapter.test.js
@@ -79,10 +79,9 @@ describe('MongooseListAdapter', () => {
           ],
         },
       },
-      { $addFields: expect.any(Object) },
       { $match: { $and: [expect.any(Object), { title: { $eq: 'bar' } }] } },
       { $addFields: { id: '$_id' } },
-      { $project: { posts: 0 } },
+      { $project: expect.any(Object) },
     ]);
 
     await userListAdapter.itemsQuery({
@@ -103,10 +102,9 @@ describe('MongooseListAdapter', () => {
           ],
         },
       },
-      { $addFields: expect.any(Object) },
       { $match: { $or: [expect.any(Object), { title: { $eq: 'bar' } }] } },
       { $addFields: { id: '$_id' } },
-      { $project: { posts: 0 } },
+      { $project: expect.any(Object) },
     ]);
   });
 
@@ -167,10 +165,9 @@ describe('MongooseListAdapter', () => {
           ],
         },
       },
-      { $addFields: expect.any(Object) },
       { $match: expect.any(Object) },
       { $addFields: { id: '$_id' } },
-      { $project: { posts: 0 } },
+      { $project: expect.any(Object) },
     ]);
 
     await userListAdapter.itemsQuery({
@@ -191,10 +188,9 @@ describe('MongooseListAdapter', () => {
           ],
         },
       },
-      { $addFields: expect.any(Object) },
       { $match: expect.any(Object) },
       { $addFields: { id: '$_id' } },
-      { $project: { posts: 0 } },
+      { $project: expect.any(Object) },
     ]);
   });
 });