+++ packages/mongo-join-builder/lib/tokenizers.js
@@ -26,7 +26,6 @@ const getRelatedListAdapterFromQueryPath = (listAdapter, queryPath) => {
     const fieldAdapter = foundListAdapter.findFieldAdapterForQuerySegment(segment);
 
     if (!fieldAdapter) {
-      // Prettier, you're testing me. Please stop.
       // prettier-ignore
       throw new Error(
           `'${listAdapter.key}' Mongo List Adapter failed to determine field responsible for the`
@@ -40,7 +39,6 @@ const getRelatedListAdapterFromQueryPath = (listAdapter, queryPath) => {
     foundListAdapter = fieldAdapter.getRefListAdapter();
 
     if (!foundListAdapter) {
-      // Seriously, though, Prettier. Don't.
       // prettier-ignore
       throw new Error(
           `'${currentKey}' Mongo List Adapter doesn't have a related list.`
@@ -61,27 +59,31 @@ const relationshipTokenizer = (listAdapter, query, queryKey, path, uid) => {
   // TODO: warn?
   if (!fieldAdapter) return {};
   const filterType = {
-    [fieldAdapter.path]: 'every',
+    [fieldAdapter.path]: 'only',
     [`${fieldAdapter.path}_every`]: 'every',
     [`${fieldAdapter.path}_some`]: 'some',
     [`${fieldAdapter.path}_none`]: 'none',
   }[queryKey];
 
+  // We use `ifNull` here to handle the case unique to mongo where a record may be
+  // entirely missing a field (or have the value set to `null`).
+  const field = fieldAdapter.path;
+  const uniqueField = `${uid}_${field}`;
+  const fieldSize = { $size: `$${uniqueField}` };
+  const expr = {
+    only: { $eq: [fieldSize, 1] },
+    every: { $eq: [fieldSize, { $size: { $ifNull: [`$${field}`, []] } }] },
+    none: { $eq: [fieldSize, 0] },
+    some: { $gt: [fieldSize, 0] },
+  }[filterType];
+
   return {
-    // The conditions under which an item from the 'orders' collection is
-    // considered a match and included in the end result
-    // All the keys on an 'order' are available, plus 3 special keys:
-    // 1) <uid>_<field>_every - is `true` when every joined item matches the
-    //    query
-    // 2) <uid>_<field>_some - is `true` when some joined item matches the
-    //    query
-    // 3) <uid>_<field>_none - is `true` when none of the joined items match
-    //    the query
-    matchTerm: { [`${uid}_${fieldAdapter.path}_${filterType}`]: true },
+    matchTerm: { $expr: expr },
     relationshipInfo: {
       from: fieldAdapter.getRefListAdapter().model.collection.name, // the collection name to join with
       field: fieldAdapter.path, // The field on this collection
       many: fieldAdapter.field.many, // Flag this is a to-many relationship
+      uniqueField,
     },
   };
 };