+++ packages/mongo-join-builder/tests/index.test.js
@@ -1,4 +1,4 @@
-const { queryParser, pipelineBuilder, mutationBuilder } = require('../');
+const { queryParser, pipelineBuilder } = require('../');
 const { listAdapter } = require('./utils');
 
 describe('Test main export', () => {
@@ -56,57 +56,40 @@ describe('Test main export', () => {
       },
     ];
     const pipeline = pipelineBuilder(queryTree);
-    const postQueryMutations = mutationBuilder(queryTree.relationships);
 
     const aggregate = jest.fn(() => Promise.resolve(aggregateResponse));
-    const result = await aggregate(pipeline).then(postQueryMutations);
+    const result = await aggregate(pipeline);
     expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts',
           as: 'posts_every_posts',
-          let: { posts_every_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$posts_every_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
                 from: 'tags',
                 as: 'tags_some_tags',
-                let: { tags_some_tags_ids: { $ifNull: ['$tags', []] } },
+                let: { tmpVar: { $ifNull: ['$tags', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$tags_some_tags_ids'] } } },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                   { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
               },
             },
             {
-              $addFields: {
-                tags_some_tags_every: {
-                  $eq: [{ $size: '$tags_some_tags' }, { $size: { $ifNull: ['$tags', []] } }],
-                },
-                tags_some_tags_none: { $eq: [{ $size: '$tags_some_tags' }, 0] },
-                tags_some_tags_some: { $gt: [{ $size: '$tags_some_tags' }, 0] },
-              },
-            },
-            {
               $match: {
-                $and: [{ title: { $eq: 'hello' } }, { tags_some_tags_some: true }],
+                $and: [
+                  { title: { $eq: 'hello' } },
+                  { $expr: { $gt: [{ $size: '$tags_some_tags' }, 0] } },
+                ],
               },
             },
-            {
-              $addFields: { id: '$_id' },
-            },
+            { $addFields: { id: '$_id' } },
+            { $project: { tags_some_tags: 0 } },
           ],
-        },
-      },
-      {
-        $addFields: {
-          posts_every_posts_every: {
-            $eq: [{ $size: '$posts_every_posts' }, { $size: { $ifNull: ['$posts', []] } }],
-          },
-          posts_every_posts_none: { $eq: [{ $size: '$posts_every_posts' }, 0] },
-          posts_every_posts_some: { $gt: [{ $size: '$posts_every_posts' }, 0] },
         },
       },
       {
@@ -114,11 +97,16 @@ describe('Test main export', () => {
           $and: [
             { name: { $eq: 'foobar' } },
             { age: { $eq: 23 } },
-            { posts_every_posts_every: true },
+            {
+              $expr: {
+                $eq: [{ $size: '$posts_every_posts' }, { $size: { $ifNull: ['$posts', []] } }],
+              },
+            },
           ],
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { posts_every_posts: 0 } },
     ]);
 
     expect(result).toMatchObject([